#2 create MainMenu.swift file and erase every codes in main.swift
